+++ src/components/Header/RightMenu/AuthLink.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link as MaterialLink } from "@mui/material";
-import Auth from "@/components/Auth";
+import Auth from "@/modal/Auth";
 
 const AuthLink = () => {
     const [isOpened, setIsOpened] = React.useState(false);
