+++ src/modal/Cart/index.tsx
@@ -4,10 +4,10 @@ import { useTypedDispatch, useTypedSelector } from "../../hooks/redux";
 import { cartSelector } from "../../store/selectors/cart";
 import { getCartItems } from "../../store/slices/cart/cart-thunks";
 import { RequestStatus } from "../../store/slices/cart/Types";
-import HeaderWithClose from "../common/HeaderWithClose";
-import BasicPreloader from "../Loaders/BasicPreloader";
-import CartItemLoader from "../Loaders/CartItemLoader";
-import LoadingError from "../LoadingError";
+import HeaderWithClose from "../../components/common/HeaderWithClose";
+import BasicPreloader from "../../components/Loaders/BasicPreloader";
+import CartItemLoader from "../../components/Loaders/CartItemLoader";
+import LoadingError from "../../components/LoadingError";
 import CartError from "./CartError";
 import CartFooter from "./CartFooter";
 import CartIsEmpty from "./CartIsEmpty";
